Sprint 4 cd test

+++ app/db/models/project.py
@@ -9,10 +9,10 @@ if TYPE_CHECKING:
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import DateTime, ForeignKey, String, Text, func, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from app.db.session import Base
+from app.db.models.base import Base
 
 
 class Project(Base):
@@ -23,7 +23,7 @@ class Project(Base):
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
     owner_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False)
     total_size_bytes: Mapped[int] = mapped_column(
-        Integer, nullable=False, server_default="0", default=0
+        BigInteger, nullable=False, server_default="0", default=0
     )
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     updated_at: Mapped[datetime] = mapped_column(
